move everything into decimalParser, including option wrangling.
