+++ src/index.ts
@@ -11,4 +11,3 @@ export * from "./validation-props";
 export * from "./controlled";
 export * from "./state";
 export * from "./messages";
-export * from "./decimal";
